Move tests from render.rs to render/filters.rs
